fixed graph
