+++ lib/analytics/managment.ex
@@ -1,0 +1,3 @@
+defmodule Gitmetrics.Managment  do
+
+end
